Lots of new changes
Added support for replacing Dungeon Encounter ids to read from the
DungeonEncounter.db2 file to get friendly names. (Optional, disabled by
default and can be turned on from the options.)

Added support for replacing Item ids to read from the Item-sparse.db2 to
get friendly names. (Optional, disabled by default because it takes a
while to load, and not many scenarios use it. Can be turned on from the
options)